Share Connect interactive auth helpers
