+++ packages/connect/src/authorization.ts
@@ -1,4 +1,9 @@
 import { getVercelOidcToken } from '@vercel/oidc';
+import {
+  isDetachedInteractiveAuth,
+  validateCallbackUrl,
+  validateWebhookUrl,
+} from './internal/url-validation.js';
 import type { ConnectTokenParams } from './token.js';
 
 export interface ConnectAuthorizationOptions {
@@ -40,9 +45,6 @@ export interface ConnectAuthorizationResponse {
   };
 }
 
-const DETACHED_INTERACTIVE_AUTH_MODE = 'detached';
-const INTERACTIVE_AUTH_MODE_ENV = 'VERCEL_CONNECT_INTERACTIVE_AUTH_MODE';
-
 export async function startAuthorization(
   connector: string,
   params: ConnectTokenParams,
@@ -51,8 +53,7 @@ export async function startAuthorization(
   if (!connector) {
     throw new Error('connector is required');
   }
-  const detachedInteractiveAuth =
-    process.env[INTERACTIVE_AUTH_MODE_ENV] === DETACHED_INTERACTIVE_AUTH_MODE;
+  const detachedInteractiveAuth = isDetachedInteractiveAuth();
   if (!detachedInteractiveAuth && options?.callbackUrl !== undefined) {
     validateCallbackUrl(options.callbackUrl);
   }
@@ -103,40 +104,4 @@ export async function startAuthorization(
 
   const data: ConnectAuthorizationResponse = await response.json();
   return data;
-}
-
-function validateCallbackUrl(value: string): void {
-  let url: URL;
-  try {
-    url = new URL(value);
-  } catch {
-    throw new Error(`Invalid callbackUrl: ${value}`);
-  }
-  if (url.protocol === 'https:') return;
-  if (url.protocol === 'http:' && isLocalHttpCallbackHostname(url.hostname)) {
-    return;
-  }
-  throw new Error(
-    `callbackUrl must be https://, http://localhost, or http://*.localhost, got: ${value}`
-  );
-}
-
-function isLocalHttpCallbackHostname(hostname: string): boolean {
-  return (
-    hostname === 'localhost' ||
-    hostname.endsWith('.localhost') ||
-    hostname === '127.0.0.1'
-  );
-}
-
-function validateWebhookUrl(value: string): void {
-  let url: URL;
-  try {
-    url = new URL(value);
-  } catch {
-    throw new Error(`Invalid webhook URL: ${value}`);
-  }
-  if (url.protocol !== 'https:') {
-    throw new Error(`webhook must be https://, got: ${value}`);
-  }
 }
